Mark Java 6 wrapper tests as not supporting non-ascii paths
